Accepts GitLab's embedded Git non-standard version suffix
The GitLab server [0] comes with an "embedded git" binary which is used
by its Git server hooks. It's version string uses a non-standard suffix
like "2.37.1.gl1". In order to accept this and potentially other Git
server vendors the looks_like_git function is being relaxed so that it
accepts any sequence of one or more alphabetic characters in place of
the standard "rc" (for release candidates) suffix.

Ref0: https://gitlab.com/rluna-gitlab/gitlab-ce